--- src/app/requests/requests.module.css
@@ -1,377 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300..800;1,300..800&display=swap');
-
-.parent-container{
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    margin-top: 100px;
-    /* border: 1px solid black; */
-}
-
-.wrapper {
-    position: relative;
-    display: flex;
-    justify-content: center;
-    flex-direction: row;
-    gap: 10px;
-    flex-wrap: wrap;
-  }
-  
-  .card {
-    position: relative;
-    width: 400px;
-    max-width: 400px;
-    height: 100px;
-    background: #fff;
-    border-radius: 10px;
-    transition: all 0.3s;
-  }
-
-  
-  
-  .card:hover {
-    transform: scale(1.05);
-  }
-  
-  .input {
-    position: relative;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    cursor: pointer;
-    appearance: none;
-    border: 1px solid #e5e5e5;
-    border-radius: 10px;
-    z-index: 10;
-    box-shadow: 1px 1px 10px #aaaaaa,
-                -1px -1px 10px #ffffff;
-  }
-  
-  .input + .check::before {
-    content: "";
-    position: absolute;
-    top: 15px;
-    right: 15px;
-    width: 16px;
-    height: 16px;
-    border: 2px solid #d0d0d0;
-    border-radius: 50%;
-    background-color: #E8E8E8;
-  }
-  
-  .input:checked + .check::after {
-    content: '';
-    position: absolute;
-    top: 19px;
-    right: 19px;
-    width: 12px;
-    height: 12px;
-    background-color: rgba(255,0,0,0.7);
-    border-radius: 50%;
-  }
-  
-  .input[value="standart"]:checked + .check::after {
-    background-color: rgba(255,165,0,0.7);
-  }
-  
-  .input[value="premium"]:checked + .check::after {
-    background-color: rgba(0,128,0,0.7);
-  }
-  
-  .input[value="basic"]:checked {
-    border: 1.5px solid rgba(255,0,0,0.7);
-  }
-  
-  .input[value="standart"]:checked {
-    border: 1.5px solid rgba(255,165,0,0.7);
-  }
-  
-  .input[value="premium"]:checked {
-    border: 1.5px solid rgba(0,128,0,0.7);
-  }
-  
-  .label {
-    color: #323232;
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 0;
-  }
-  
-  .label .title {
-    margin: 15px 0 0 15px;
-    font-weight: 900;
-    font-size: 15px;
-    letter-spacing: 1.5px;
-  }
-  
-  .label .price {
-    margin: 20px 0 0 15px;
-    font-size: 10px;
-    font-weight: 900;
-  }
-  
-  .label .span {
-    color: gray;
-    font-weight: 700;
-    font-size: 15px;
-  }
-  
-
-  .raise-request-container{
-    display: flex;
-    justify-content: center;
-  }
-  .lower_card {
-    padding: 20px;
-    width: 400px;
-    display: flex;
-    justify-content: center;
-    /* background: #FFFFFF; */
-    /* box-shadow: 0px 187px 75px rgba(0, 0, 0, 0.01), 0px 105px 63px rgba(0, 0, 0, 0.05), 0px 47px 47px rgba(0, 0, 0, 0.09), 0px 12px 26px rgba(0, 0, 0, 0.1), 0px 0px 0px rgba(0, 0, 0, 0.1); */
-  }
-
-  .my_title {
-    width: 100%;
-    height: 40px;
-    position: relative;
-    display: flex;
-    align-items: center;
-    padding-left: 20px;
-    border-bottom: 1px solid #efeff3;
-    font-weight: 700;
-    font-size: 20px;
-    color: #63656b;
-  }
-  
-  /* coupons */
-.coupons {
-    border-radius: 7px;
-  }
-  
-  .coupons form {
-    display: flex;
-    grid-template-columns: 1fr 80px;
-    gap: 10px;
-    padding: 10px;
-  }
-  
-  .input_field {
-    width: auto;
-    height: 36px;
-    padding: 0 0 0 12px;
-    border-radius: 5px;
-    outline: none;
-    border: 1px solid #e5e5e5;
-    filter: drop-shadow(0px 1px 0px #efefef)
-      drop-shadow(0px 1px 0.5px rgba(239, 239, 239, 0.5));
-    transition: all 0.3s cubic-bezier(0.15, 0.83, 0.66, 1);
-  }
-  
-  .input_field:focus {
-    border: 1px solid transparent;
-    box-shadow: 0px 0px 0px 2px #242424;
-    background-color: transparent;
-  }
-  
-  .btn {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding: 10px 18px;
-    gap: 10px;
-    width: 100%;
-    height: 36px;
-    background: linear-gradient(180deg, #4480FF 0%, #115DFC 50%, #0550ED 100%);
-    box-shadow: 0px 0.5px 0.5px #EFEFEF, 0px 1px 0.5px rgba(239, 239, 239, 0.5);
-    border-radius: 5px;
-    border: 0;
-    font-style: normal;
-    font-weight: 600;
-    font-size: 12px;
-    line-height: 15px;
-    color: #ffffff;
-  }
-/* 
-  .order_title{
-
-    
-  }
-  .order_status{
-
-
-  }
-  .order_date{
-
-
-  }
-
-  .order_headings{
-    margin-top: 30px;
-    display: flex;
-    gap: 100px;
-    font-weight: 600;
-    font-size: 2em;
-  }
-
-  .request_order{
-display: flex;
-flex-direction: column;
-
-  }
-
-  .order_container{
-    margin-top: 20px;
-    display: flex;
-    gap: 150px;
-  } */
-
-  /* ResponsiveTable.module.css */
-.order_headings {
-    display: flex;
-    justify-content: space-between;
-    font-weight: bold;
-    padding: 10px;
-    background-color: #f4f4f4;
-    border-bottom: 2px solid #ddd;
-  }
-  
-  .request_order {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-  }
-  
-  .order_container {
-    display: flex;
-    justify-content: space-between;
-    padding: 10px;
-    border-bottom: 1px solid #ddd;
-    align-items: center;
-  }
-  
-  .order_title, .order_date, .order_status {
-    width: 33%;
-    text-align: center;
-  }
-  
-  /* Responsiveness for smaller screens */
-  @media (max-width: 768px) {
-    .order_headings, .order_container {
-      flex-direction: column;
-      text-align: left;
-    }
-  
-    .order_title, .order_date, .order_status {
-      width: 100%;
-      margin-bottom: 5px;
-    }
-  
-    .order_container {
-      border-bottom: 1px solid #ddd;
-      padding: 15px 0;
-    }
-  }
-  
-
-  /* General
-   Table Styling */
-.table {
-    font-family: "Open Sans", sans-serif;
-    margin-top: 30px;
-    border: 1px solid #ccc;
-    border-collapse: collapse;
-    /* margin: 0; */
-    padding: 0;
-    width: 100%;
-    table-layout: fixed;
-  }
-  
-  .caption {
-    font-size: 1.5em;
-    margin: 0.5em 0 0.75em;
-  }
-  
-  .tr {
-    background-color: #f8f8f8;
-    border: 1px solid #ddd;
-    padding: 0.35em;
-  }
-  
-  .th,
-  .td {
-    padding: 0.625em;
-    text-align: center;
-  }
-  
-  .th {
-    font-size: 0.85em;
-    letter-spacing: 0.1em;
-    text-transform: uppercase;
-  }
-  
-  /* Responsive Styles for Mobile */
-  @media screen and (max-width: 600px) {
-    .table {
-      border: 0;
-    }
-  
-    .caption {
-      font-size: 1.3em;
-    }
-  
-    .thead {
-      border: none;
-      clip: rect(0 0 0 0);
-      height: 1px;
-      margin: -1px;
-      overflow: hidden;
-      padding: 0;
-      position: absolute;
-      width: 1px;
-    }
-  
-    .tr {
-      border-bottom: 3px solid #ddd;
-      display: block;
-      margin-bottom: 0.625em;
-    }
-  
-    .td {
-      border-bottom: 1px solid #ddd;
-      display: block;
-      font-size: 0.8em;
-      text-align: right;
-    }
-  
-    .td::before {
-      content: attr(data-label);
-      float: left;
-      font-weight: bold;
-      text-transform: uppercase;
-    }
-  
-    .td:last-child {
-      border-bottom: 0;
-    }
-  }
-  
-
-.request_logo{
-
-    width: 18px;
-    margin-right: 5px;
-    margin-top: 5px;
-}
-
-.ststyle{
-    display: inline;
-  color: white; /* Set text color */
-  padding: 2px 10px; 
-  border-radius: 30px;
-  font-weight: 300;
-  font-size: 0.8em;
-}
